feat: Calculator 클래스 생성 (사칙연산 수행 후 결과값 반환 메서드와 연산 결과를 저장 필드를 가짐)
